project_fifth_term
We developed a project involving a SISO (Shift-In, Shift-Out) register that is converted into a SIPO (Shift-In, Parallel-Out) register using a switch. 🔄 The project also controls data in two ways. The first method uses a NE555 timer, while the second method allows manual data input with a pulse given to the clock by me. ⏱️